requirements.txt
#NOTE: plugin fails to load if these packages are not pre-installed, automated install not working
ruamel.yaml
requests
